+++ library/core/src/main/java/com/google/android/exoplayer2/ExoPlayerLibraryInfo.java
@@ -51,6 +51,9 @@ public final class ExoPlayerLibraryInfo {
    */
   public static final boolean ASSERTIONS_ENABLED = true;
 
+  /** Whether an exception should be thrown in case of an OpenGl error. */
+  public static final boolean GL_ASSERTIONS_ENABLED = false;
+
   /**
    * Whether the library was compiled with {@link com.google.android.exoplayer2.util.TraceUtil}
    * trace enabled.